refactor: create overload groups in non-unified symbol graphs
