+++ Sources/SymbolKit/UnifiedSymbolGraph/UnifiedSymbolGraph.swift
@@ -177,82 +177,6 @@ extension UnifiedSymbolGraph {
 
         loadRelationships(fromGraph: graph)
     }
-
-    /// Iterates through the symbols in this graph and inserts new "overload group" symbols for
-    /// any overloaded symbols found.
-    ///
-    /// An "overload" here is defined as any symbol whose path components and kind conflicts with
-    /// another symbol. This method will collect all such conflicting symbols together and create a
-    /// new "overload group" symbol, cloned from an arbitrarily-selected "first overload" and given
-    /// a new unique identifier.
-    internal func createOverloadGroups(restrictByLanguage languages: [String]? = nil) {
-        struct OverloadKey: Hashable {
-            let path: [String]
-            let kind: SymbolGraph.Symbol.KindIdentifier
-        }
-
-        var symbolsByPath = [Selector: [OverloadKey: [UnifiedSymbolGraph.Symbol]]]()
-
-        for symbol in self.symbols.values {
-            for selector in symbol.mainGraphSelectors {
-                // If we've been given a list of languages to restrict the collection with,
-                // double-check that this selector is actually for one of those languages
-                if let languages = languages, !languages.contains(selector.interfaceLanguage) {
-                    continue
-                }
-                if let path = symbol.pathComponents[selector], let kind = symbol.kind[selector], kind.identifier.isOverloadableKind {
-                    symbolsByPath[selector, default: .init()][.init(path: path, kind: kind.identifier), default: []].append(symbol)
-                }
-            }
-        }
-
-        for (selector, symbols) in symbolsByPath {
-            var newRelationships = [SymbolGraph.Relationship]()
-
-            for overloads in symbols.values where overloads.count > 1 {
-                let firstOverload = overloads.first!
-
-                // Create a new "overload group" symbol and add it to the list
-                let overloadGroupIdentifier = firstOverload.uniqueIdentifier + "::OverloadGroup"
-                let overloadGroupSymbol = UnifiedSymbolGraph.Symbol(cloning: firstOverload, withIdentifier: overloadGroupIdentifier)
-
-                if let simplifiedDeclaration = overloadGroupSymbol.overloadSubheadingFragments(selector: selector),
-                   var overloadNames = overloadGroupSymbol.names[selector]
-                {
-                    overloadNames.navigator = simplifiedDeclaration
-                    overloadNames.subHeading = simplifiedDeclaration
-                    overloadGroupSymbol.names[selector] = overloadNames
-                }
-
-                self.symbols[overloadGroupIdentifier] = overloadGroupSymbol
-
-                // Clone the relationships from the first overload and add them to the overload group
-                for relationship in self.relationshipsByLanguage[selector] ?? [] where relationship.source == firstOverload.uniqueIdentifier {
-                    var newRelationship = relationship
-                    newRelationship.source = overloadGroupIdentifier
-                    newRelationships.append(newRelationship)
-                }
-                for relationship in self.relationshipsByLanguage[selector] ?? [] where relationship.target == firstOverload.uniqueIdentifier {
-                    var newRelationship = relationship
-                    newRelationship.target = overloadGroupIdentifier
-                    newRelationships.append(newRelationship)
-                }
-
-                // Make new 'overloadOf' relationships between the overloaded symbols and the new overload group
-                for overloadSymbol in overloads {
-                    newRelationships.append(.init(
-                        source: overloadSymbol.uniqueIdentifier,
-                        target: overloadGroupIdentifier,
-                        kind: .overloadOf,
-                        targetFallback: nil))
-                }
-            }
-
-            if newRelationships.count > 0 {
-                self.relationshipsByLanguage[selector, default: []].append(contentsOf: newRelationships)
-            }
-        }
-    }
 }
 
 extension UnifiedSymbolGraph {
